+++ spec/controllers/paper_conversions_controller_spec.rb
@@ -3,9 +3,15 @@ require 'rails_helper'
 describe PaperConversionsController, type: :controller do
   include Rails.application.routes.url_helpers
 
-  let(:paper) { FactoryGirl.create(:paper) }
+  let(:paper) { instance_double(Paper, id: 99, file: nil, to_param: '99') }
   let(:job_id) { 'd5ee706f-a473-46ed-9777-3b7cd2905d08' }
   let(:user) { FactoryGirl.create :user }
+
+  before do
+    allow(Paper).to receive(:find)
+      .with(paper.to_param)
+      .and_return paper
+  end
 
   describe 'GET export' do
     subject(:do_request) do
@@ -15,14 +21,15 @@ describe PaperConversionsController, type: :controller do
     it_behaves_like 'an unauthenticated json request'
 
     context 'as a user with access to a paper' do
+      let(:manuscript_attachment) do
+        FactoryGirl.build_stubbed(:manuscript_attachment)
+      end
+
       before do
         stub_sign_in(user)
         allow(user).to receive(:can?)
           .with(:view, paper)
           .and_return true
-        allow(Paper).to receive(:find)
-          .with(paper.to_param)
-          .and_return paper
       end
 
       context 'with a paper that needs conversion' do
@@ -34,8 +41,8 @@ describe PaperConversionsController, type: :controller do
 
         before do
           # no source URL, needs conversion
-          allow(paper).to receive_message_chain('latest_version.source_url')
-            .and_return nil
+          allow(paper).to receive(:file).and_return manuscript_attachment
+          expect(paper.file.url).to be(nil)
 
           allow(PaperConverter).to receive(:export)
             .and_return double('job', job_id: job_id)
@@ -58,11 +65,12 @@ describe PaperConversionsController, type: :controller do
       end
 
       context 'with a docx that the user uploaded' do
-        let(:docx_url) { 'http://example.com/source.docx' }
+        let(:manuscript_attachment) do
+          FactoryGirl.build_stubbed(:manuscript_attachment, :with_filename)
+        end
 
         before do
-          allow(paper).to receive_message_chain('latest_version.source_url')
-            .and_return docx_url
+          allow(paper).to receive(:file).and_return manuscript_attachment
         end
 
         it 'returns a url to check later' do
@@ -100,18 +108,17 @@ describe PaperConversionsController, type: :controller do
     it_behaves_like 'an unauthenticated json request'
 
     context "when the user has access" do
-      let(:docx_url) { 'http://example.com/source.docx' }
+      let(:manuscript_attachment) do
+        FactoryGirl.build_stubbed(:manuscript_attachment, :with_filename)
+      end
 
       before do
         stub_sign_in(user)
         allow(user).to receive(:can?)
           .with(:view, paper)
           .and_return true
-        allow(Paper).to receive(:find)
-          .with(paper.to_param)
-          .and_return paper
-        allow(paper).to receive_message_chain('latest_version.source_url')
-          .and_return docx_url
+        allow(paper).to receive(:file).and_return manuscript_attachment
+        expect(paper.file.url).to_not be(nil)
       end
 
       it 'returns 202 when still processing' do
@@ -124,7 +131,7 @@ describe PaperConversionsController, type: :controller do
       it 'returns the download url when the status is checked' do
         get :status, id: paper.id, job_id: 'source', export_format: 'docx', format: :json
         expect(response.status).to eq(200)
-        expect(res_body['url']).to eq(docx_url)
+        expect(res_body['url']).to eq(manuscript_attachment.url)
       end
     end