+++ spec/factories/manuscript_attachment_factory.rb
@@ -1,0 +1,12 @@
+FactoryGirl.define do
+  factory :manuscript_attachment, parent: :attachment, class: 'ManuscriptAttachment' do
+    association :owner, factory: :paper
+    paper
+
+    trait :with_filename do
+      after :stub do |manuscript_attachment|
+        manuscript_attachment[:file] = 'source.docx'
+      end
+    end
+  end
+end