+++ spec/models/invitation_attachment_spec.rb
@@ -1,0 +1,20 @@
+require 'rails_helper'
+
+describe InvitationAttachment do
+  subject(:attachment) do
+    FactoryGirl.create(:invitation_attachment)
+  end
+
+  describe '#download!', vcr: { cassette_name: 'attachment' } do
+    let(:url) { 'http://tahi-test.s3.amazonaws.com/temp/bill_ted1.jpg' }
+
+    it_behaves_like 'attachment#download! raises exception when it fails'
+    it_behaves_like 'attachment#download! stores the file'
+    it_behaves_like 'attachment#download! caches the s3 store_dir'
+    it_behaves_like 'attachment#download! sets the file_hash'
+    it_behaves_like 'attachment#download! sets title to file name'
+    it_behaves_like 'attachment#download! sets the status'
+    it_behaves_like 'attachment#download! always keeps snapshotted files on s3'
+    it_behaves_like 'attachment#download! manages resource tokens'
+  end
+end