+++ spec/workers/download_manuscript_worker_spec.rb
@@ -1,34 +1,48 @@
 require 'rails_helper'
 
 describe DownloadManuscriptWorker, redis: true do
-  let(:paper) do
-    FactoryGirl.create(:paper, :with_integration_journal, :with_creator)
-  end
-  let(:url) { 'https://example.com/temp/about_equations.docx' }
+  let(:paper) { FactoryGirl.build_stubbed(:paper) }
+  let(:user) { FactoryGirl.build_stubbed(:user) }
+  let(:url) { 'http://tahi.example.com/paper.docx' }
 
-  before do
-    VCR.turn_off!
-    versioned_text_id = paper.latest_version.id
-    s3_url = "https://tahi-test.s3-us-west-1.amazonaws.com/uploads/\
-versioned_text/#{versioned_text_id}/about_equations.docx"
-    @docx_req = stub_request(:get, url).to_return(body: 'foo')
-    @s3_req = stub_request(:get, s3_url)
-      .with(query: hash_including)
-      .to_return(body: 'foo')
-    @ihat_status_req = stub_request(:post, 'http://ihat.example.com/jobs')
-      .with(body: /.*/)
-      .to_return(body: { job: { state: 'processing', options: {} } }.to_json)
+  describe '.download_manuscript' do
+    before do
+      allow(paper).to receive(:update_attribute)
+    end
+
+    it 'queues up a job to download the manuscript' do
+      expect(DownloadManuscriptWorker).to receive(:perform_async).with(paper.id, url, user.id)
+      described_class.download_manuscript(paper, url, user)
+    end
+
+    context 'when the URL is blank' do
+      it 'raises an exception' do
+        expect do
+          described_class.download_manuscript(paper, nil, user)
+        end.to raise_error(ArgumentError, "Url must be provided (received a blank value)")
+      end
+    end
+
+    it 'marks the paper as processing' do
+      expect(paper).to receive(:update_attribute).with(:processing, true)
+      described_class.download_manuscript(paper, url, user)
+    end
   end
 
-  after do
-    VCR.turn_on!
-  end
+  describe '#perform' do
+    before do
+      allow(User).to receive(:find).with(user.id).and_return user
+      allow(Paper).to receive(:find).with(paper.id).and_return paper
+    end
 
-  it 'downloads the attachment' do
-    DownloadManuscriptWorker.new.perform(paper.id, url, 'http://localhost/callback',
-                                         foo: 'bar')
-    expect(@docx_req).to have_been_requested
-    expect(@s3_req).to have_been_requested
-    expect(@ihat_status_req).to have_been_made.at_least_once
+    it 'downloads a manuscript with a user' do
+      expect(paper).to receive(:download_manuscript!).with(url, uploaded_by: user)
+      described_class.new.perform(paper.id, url, user.id)
+    end
+
+    it 'downloads a manuscript with a blank user' do
+      expect(paper).to receive(:download_manuscript!).with(url, uploaded_by: nil)
+      described_class.new.perform(paper.id, url, nil)
+    end
   end
 end